% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_utils.R
\name{accuracy}
\alias{accuracy}
\title{Calculate the Accuracy of the model}
\usage{
accuracy(new_Y, p, verb = FALSE)
}
\arguments{
\item{p}{Vector of fitted probabilities for each Y.}
}
\value{
The accuracy of the model, given the fitted probabilities and new data
}
\description{
Given a set of true values and their corresponding fitted probabilities,
the function calculates the accuracy of the model defined by:
 \eqn{1- #wrong prediction/# of observations}
}
\examples{
(new_Y, fitted_probs_for_data)
}
