% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpwpm2.R
\name{bpwpm2_gibbs}
\alias{bpwpm2_gibbs}
\title{Bayesian Piece-Wise Polynomial Model 2 Algorithm}
\usage{
bpwpm2_gibbs(Y, X, M, J, K, draws = 10^3, tau = NULL, beta_init = NULL,
  mu_beta_0 = NULL, sigma_beta_0_inv = NULL, precision_beta = 1,
  eps = 1e-15, verb = FALSE, debug_verb = FALSE)
}
\arguments{
\item{Y}{Response vector of n binary observatios (integers 0,1 - vector of
size n) Can be encoded as a factor a numeric vector.}

\item{X}{Design matrix of n observations and d covariables (numeric - n*d)}

\item{M}{M minus 1 is the degree of the polinomial (integer - M > 0)}

\item{J}{Number of intervals in each dimention (integer - J > 1)}

\item{K}{Order of continuity in the derivatives (integrer - 0 < K < M)}

\item{draws}{Númber of samples to draw from the Gibbs Sampler (integer - draw
> 0)}

\item{tau}{the initial position of nodes selected by the user. although·
arbitraty they need to match the dimentions. (numeric - (J-1)*d)}

\item{beta_init}{Inital value for the Gibbs Sampler Chain (numeric - vector of
size 1*(1 + Nd))}

\item{mu_beta_0}{Prior Mean of Beta (numeric - matrix of size N*d)}

\item{sigma_beta_0_inv}{sigma_w_0_inv:= Prior Inverse if the Variance-Covariance
Matrices of w (list - d elements, each element is a numeric matrix of size
N*N)}

\item{eps}{Numerical threshold}

\item{verb}{short for verbose, if TRUE, prints aditional information (logical)}

\item{debug_verb}{If TRUE, print even more info to help with debug_verbging (logical)}

\item{precision_w}{If using the default sigmas for w, a diagonal matrix will
be used. Precision controls its magnitude (numeric - precision > 0)}
}
\value{
An object of the class "bpwpm" containing at the following
components:
\describe{
\item{beta: }{A data frame containing the Gibbs sampler simulation for beta}
\item{Psi: }{The PWP Expansion for input matrix X and nodes selected on
percentiles}
\item{tau: }{Nodes used for training}
\item{M: }{Initial parameters}
\item{J: }{Initial parameters}
\item{K: }{Initial parameters}
\item{d: }{Number of dimentions}
\item{indep_terms: }{Logical. If independent terms are keept}
\item{info}{A string that prints the basic information of the mode. Used for
the summary function.}
}
}
\description{
Second take at the bpwpm model, based solely on the Albert + Chibb Paper
}
\examples{
See the main document of thesis for a couple of full examples
with its corresponding analysis.
}
