% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_utils.R
\name{calculate_eta}
\alias{calculate_eta}
\title{Calculate prediction eta}
\usage{
calculate_eta(F_mat)
}
\arguments{
\item{F_mat}{The PWP transformed input space, calculated by
\code{\link{calculate_F}}}
}
\value{
A numeric vector representing the eta of \code{R^d} into
\code{R} given all of the parametres
}
\description{
Calculate prediction eta
}
