% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_utils.R
\name{calculate_F}
\alias{calculate_F}
\title{F matrix calculation}
\usage{
calculate_F(Psi, beta, d)
}
\arguments{
\item{Psi}{Piecewise Polinomail expansion for an input matrix X previosly
calculated by \code{\link{calculate_Psi}}}

\item{beta}{Posterior estimate of the beta parameters previously calculated
with \code{\link{posterior_params}}}

\item{d}{Number of dimentions, this parameter helps to improve efficiency}
}
\value{
F matrix
}
\description{
Function for calculating the F matrix, described on the thesis.
This is the transformed input matrix that depends on the piecewise polinomial
expansion Psi and a set of weights w.
}
