% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_utils.R
\name{calculate_Psi}
\alias{calculate_Psi}
\title{Piece wise polinomial expansion for X (PWP)}
\usage{
calculate_Psi(X, M, J, K, n, d, tau)
}
\arguments{
\item{X}{Design matrix of n observations and d covariables (numeric - n*d)}

\item{M}{M minus 1 is the degree of the polinomial (integer - M > 0)}

\item{J}{Number of intervals in each dimention (integer - J > 1)}

\item{K}{Order of continuity in the derivatives (integrer - 0 < K < M)}

\item{tau}{}
}
\value{
A desginer matrix containing the PWP expansion described in (...)
}
\description{
Calculates and returns a new designer matrix, representing the correspondieng PWP
expansion for all d variables. Combines all the parameters on a relatively
fast computation of basis expansion for X described on the thesis and on
(Denison, Mallik and Smith, 1998).
}
