% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_utils.R
\name{contingency_table}
\alias{contingency_table}
\title{Contingency Table for the prediciton of a bpwpm}
\usage{
contingency_table(new_Y, p)
}
\arguments{
\item{new_Y}{Response variable to test the model for}

\item{p}{Vector of fitted probabilities}
}
\value{
The contingency table
}
\description{
Contingency Table for the prediciton of a bpwpm
}
\examples{
contingency_table(train_y, est_p)
}
