% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_utils.R
\name{ergodic_mean}
\alias{ergodic_mean}
\title{Ergodic Mean}
\usage{
ergodic_mean(mcmc_chain)
}
\arguments{
\item{mcmc_chain}{}
}
\value{
The Ergodic Mean Matrix
}
\description{
Calculates the Ergodic Mean for an MCMC_Chain Matrix producede as output from
\code{\link{bpwpm_chain}}. It is used by the \code{\link{plot_ergodic_mean}}
 function, but left available for the user.
}
\examples{
MA_betas <- (betas)
}
