% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_utils.R
\name{log_loss}
\alias{log_loss}
\title{Log Loss}
\usage{
log_loss(Y, p, eps = 1e-10, verb = TRUE)
}
\arguments{
\item{p}{Vector of fitted probabilities for each Y.}

\item{eps}{Machine error to hanlde limit cases on the logarithm function}
}
\value{
The value of the Log Loss function (numeric). The smaller, the better
}
\description{
An implementation of the Log-Loss function for the binomial case
}
\examples{
log_loss(true_values, fitted probabilities)
log_loss(true_values, fitted probabilities)
log_loss(true_values, fitted probabilities, 1e-30, FALSE)
}
