% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.R
\name{plot.bpwpm}
\alias{plot.bpwpm}
\title{Generic bpwpm object plotting}
\usage{
\method{plot}{bpwpm}(object, n = 1000, m = 5, ...)
}
\arguments{
\item{object}{of the class bpwpm}

\item{n}{number of draws to plot (the function plots from the end, ie:
n = 100 will plot the last 100 draws of the chain)}

\item{...}{additional parameters to be passed to the functions}
}
\value{
a series of line plots and histograms
}
\description{
Once a bpwpm has been run using the function \code{\link{bpwpm2_gibbs}}, the
chains can be plotted and hence evaluated. This generic function creates 3
set of graphs for the parameters \eqn{\beta} of the model:
- Ergodic means
- Trace of the parameters (in groups of size m)
- Histograms of the parameters (in groups of size m)
}
\examples{
(model1, 1000), (model2, 2000)
}
