% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.R
\name{plot.bpwpm_prediction}
\alias{plot.bpwpm_prediction}
\title{Generic function for plotting bpwpm_predictions objects}
\usage{
\method{plot}{bpwpm_prediction}(object, ...)
}
\arguments{
\item{object}{of the class bpwpm_prediction}

\item{...}{other arguments}
}
\value{
a series of plots from ggplot2
}
\description{
Once a model has been run and evaluated, a prediction can be made using the
function \code{\link{predict.bpwpm}}. The input matrix \code{X} and
output \code{Y} are saved and can be plotted against the final PWP expansion
for the model.
}
\examples{
(train_set_prediciton),
 (test_set_prediciton)
}
