% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.R
\name{plot_2D_F}
\alias{plot_2D_F}
\title{Plots the transformed input space F}
\usage{
plot_2D_F(Y, bpwpm_params)
}
\arguments{
\item{Y}{A response vector of size n}

\item{bpwpm_params}{an object of the class bpwpm_params of bpwpm_prediction
created by the functions \code{\link{posterior_params}} or
\code{\link{predict.bpwpm}} respectively that contains all the info about the
posterior parametres of the model.}
}
\value{
A plot of F1(X1) and F2(X2)
}
\description{
Plots the transformed input space F
}
