% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.R
\name{plot_3D_proj}
\alias{plot_3D_proj}
\title{Plots the 3D representation of the projection function}
\usage{
plot_3D_proj(X, bpwpm_params, m, fof0 = TRUE)
}
\arguments{
\item{X}{An input matrix of size n*2.}

\item{bpwpm_params}{an object of the class bpwpm_params of bpwpm_prediction
created by the functions \code{\link{posterior_params}} or
\code{\link{predict.bpwpm}} respectively that contains all the info about the
posterior parametres of the model.}

\item{m}{Thinness of grid for 3D projection}

\item{fof0}{logical - if the constant function 0 is to be plotted}
}
\value{
a 3D WireFrame Lattice Plot
}
\description{
Given the set of parmeters and the input data in 2D, this function calculates
and plots the wireframe on a 3D linear space defined by the input matrix X.
}
