% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.R
\name{plot_each_F}
\alias{plot_each_F}
\title{Plots each dimention f(x)}
\usage{
plot_each_F(Y, X, F_mat)
}
\arguments{
\item{Y}{A vector of binary response. Can be encoded as either a factor
vector or as a numeric one.}

\item{X}{A data frame or matrix containing the original Inputs for the model.}

\item{F_mat}{The F matrix calculated via \code{\link{calculate_F}} or
alternativly you can pass it the parameters calculated by function
\code{\link{posterior_params}}}
}
\value{
d plots for each dimention created using ggplot2
}
\description{
With the posterior \code{w} parameters calculated from the Gibbs run, and
\code{\link{posterior_params}}, a Final F matrix can be calculated. and
hence, ploted against every Input X to see how does the PWP expansion looks
like for the specified set of parameters.
}
