% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.R
\name{plot_hist}
\alias{plot_hist}
\title{Plot MCMC Chains histograms}
\usage{
plot_hist(mcmc_chain, number = 100, title = "", ...)
}
\arguments{
\item{mcmc_chain}{An MCMC Chain matrix. (draws * number of params)}

\item{title}{Title for the plot}

\item{...}{additional parameters to be passed to the functions}
}
\value{
A histogram for the n draws and parameters of the chain
}
\description{
Plot the parameters to test for convergence
}
