% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{posterior_params}
\alias{posterior_params}
\title{Puntual Posterior Estimation of bpwpm2 beta parameters}
\usage{
posterior_params(bpwpm, burn_in, thin, type = "mean")
}
\arguments{
\item{bpwpm}{an object of the class bpwpm}

\item{burn_in}{A burn in parameter for the MCMC Chain}

\item{thin}{A thinning parameter for the MCMC Chain}

\item{type}{The type of punctual estimation for the parameters. Options
include: mean, mode or median.}
}
\value{
An object of the class bpwpm_params that contains:
\describe{
  \item{\eqn{\beta}}{The posterior estimation for \eqn{\beta}}
  \item{\eqn{\tau}}{The corresponding nodes}
  \item{F}{The final F matrix}
  \item{M}{M parameter}
  \item{J}{J parameter}
  \item{K}{K parameter}
  \item{d}{number of dimentions}
  \item{indep_terms}{Logical}
}
}
\description{
Given a model output by the function \code{\link{bpwpm_gibbs}}, it thins
down the chain and makes puntual estimation for the parameters, given a type
of estimation. This parameter object can later be used to calculate other
results.
}
\examples{
(model, 2, 2000, 'mean') (model, 0, 0, 'median')
}
