% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_funcs.R
\name{predict.bpwpm}
\alias{predict.bpwpm}
\title{Predict Method for a bpwpm}
\usage{
\method{predict}{bpwpm}(object, new_Y, new_X, burn_in = 0, thin = 0,
  type = "mean", ...)
}
\arguments{
\item{object}{An object of the class bpwpm}

\item{new_Y}{New Y data to make the prediction for. However, the train Y data
can be used to evaluate the model}

\item{new_X}{New X data to make the prediction for. However, the train X data
can be used to evaluate the model}

\item{burn_in}{A burn in parameter for the MCMC Chain}

\item{thin}{A thinning parameter for the MCMC Chain}

\item{type}{The type of punctual estimation for the parameters. Options
include: mean, mode or median.}

\item{...}{further arguments passed to or from other methods}
}
\value{
An object of the class \code{bpwpm_prediction} containing:
\describe{
  \item{Info}{A formated string that describes the basics of the model}
  \item{type}{The type of posterior probability used}
  \item{fitted_probabilities}{The fitted probabilities for the input}
  \item{bpwpm_params}{An object of the class \code{bpwpm_params}, created by
  the function \code{\link{posterior_params}}}
  \item{contingency_table}{The confusion matrix for this model, and
  prediction}
  \item{accuracy}{The accuracy of the model. This can be misleading.}
  \item{log_loss}{The log loss for the Y response and the fitted
  probabilities}
  \item{X}{The X input matrix passed down to the method. Used for plotting
  methods}
  \item{Y}{The Y input matrix passed down to the method. Used for plotting
  methods}
}
}
\description{
Predict Method for a bpwpm
}
\examples{
(model1, train_Y, train_X, 2, 1000, mean)
(model2, test_Y, test_X, 1, 0, mode)
}
