% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_funcs.R
\name{summary.bpwpm}
\alias{summary.bpwpm}
\title{Summarizes an bpwpm run}
\usage{
\method{summary}{bpwpm}(object, thin = 0, burn_in = 0,
  digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{object}{of the class bpwpm}

\item{digits}{number of significant digits for the figures}

\item{...}{additional parameters}
}
\value{
Prints out summarized information
}
\description{
Once the model has been run, soem relevant statistics are given for each one
of the parameters to evaluate its convergence and relevance.
}
