% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_funcs.R
\name{summary.bpwpm_prediction}
\alias{summary.bpwpm_prediction}
\title{Summarizes an bpwpm prediciton}
\usage{
\method{summary}{bpwpm_prediction}(object, digits = max(3L,
  getOption("digits") - 3L), verb = FALSE, ...)
}
\arguments{
\item{object}{An object of the class bpwpm_prediction}

\item{digits}{Number of significant digits for the figures}

\item{verb}{To print out additional info}

\item{...}{Arguments to be passed to or from other methods}
}
\value{
Prints out info
}
\description{
Summarizes an bpwpm prediciton
}
