% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{thin_bpwpm}
\alias{thin_bpwpm}
\title{Thinning of a BPWPM}
\usage{
thin_bpwpm(bpwpm, burn_in, thin)
}
\arguments{
\item{bpwpm}{A bpwpm object created by \code{\link{bpwpm_gibbs}}}

\item{burn_in}{A burn in parameter for the MCMC Chain}

\item{thin}{A thinning parameter for the MCMC Chain}
}
\value{
An object of the same kind of the bpwpm but thinned down
}
\description{
Thins the beta chain of an bpwpm2 and returns an object of the same
type
}
