% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{thin_chain}
\alias{thin_chain}
\title{Thin MCMC Chain}
\usage{
thin_chain(mcmc_chain, burn_in, thin)
}
\arguments{
\item{mcmc_chain}{An MCMC Chain matrix. (draws * number of params)}

\item{burn_in}{A burn in parameter for the MCMC Chain}

\item{thin}{A thinning parameter for the MCMC Chain}
}
\value{
The thinned down version of the MCMC chain
}
\description{
Thins and eliminates the burn in period of the Gibbs Chain.
}
\examples{
(beta, 2, 2000)
}
