% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.R
\name{plot_2D_data}
\alias{plot_2D_data}
\title{Scatter Plot of 2D data}
\usage{
plot_2D_data(Y, X)
}
\value{
A ggplot2 scatter plot
}
\description{
Scatter plot to visualize the data and it's corresponding groups.
}
\examples{
(Y = rbinom(100, 1, 4), X = cbind(rnorm(100), rnorm(100)))
}
